Clean up file structure, summary file
